% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{checksums_str}
\alias{checksums_str}
\title{String representations of the checksums and filenames}
\usage{
checksums_str(x)
}
\arguments{
\item{param}{TODO}
}
\description{
Sorted list of checksums for each file. Hash this list
}
\examples{
example
}
