% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{data_integrity_fingerprint}
\alias{data_integrity_fingerprint}
\title{Create a Data Integrity Fingerprint}
\usage{
data_integrity_fingerprint(folder, hash_algorithm = "sha256")
}
\arguments{
\item{param}{TODO}
}
\description{
Create a Data Integrity Fingerprint
}
\examples{
example
}
