% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{file_hash}
\alias{file_hash}
\title{Create a checksum for a given file.}
\usage{
file_hash(filename, hash_algorithm = "sha256")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Create a checksum for a given file.
}
\examples{
file_hash()
}
