% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{load_checksums}
\alias{load_checksums}
\title{TODO A Cat Function}
\usage{
load_checksums(filename, hash_algorithm)
}
\arguments{
\item{param}{TODO}
}
\description{
This function allows you to express your love of cats.
}
\examples{
example
}
