% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{master_hash}
\alias{master_hash}
\title{TODO A Cat Function}
\usage{
master_hash(x)
}
\arguments{
\item{param}{TODO}
}
\description{
This function allows you to express your love of cats.
}
\examples{
example
}
