% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{sort.DIF}
\alias{sort.DIF}
\title{Sort checksums of a data integrity fingerprint}
\usage{
\method{sort}{DIF}(x)
}
\arguments{
\item{param}{TODO}
}
\description{
Sort checksums of a data integrity fingerprint
}
\examples{
example
}
